+++ packages/extra/fs/src/rm.ts
@@ -11,21 +11,32 @@ import {
 } from 'fs'
 
 export interface RmOptions extends RmDirOptions {
+	/**
+	 * force delete, default: true
+	 */
 	force?: boolean
+	/**
+	 * do not output logs, default: true
+	 */
+	silent?: boolean
 }
 
 /**
  * remove file
  *
  * @param paths - path or file, support array
+ * @param options - RmOptions
  */
-export async function rm(paths: PathLike | PathLike[], options?: RmOptions): Promise<void> {
+export async function rm(
+	paths: PathLike | PathLike[],
+	options: RmOptions = { silent: true, force: true }
+): Promise<void> {
 	if (!(paths instanceof Array)) paths = ([] as PathLike[]).concat(paths)
 
 	for (let path of paths) {
 		typeof path === 'string' && !isAbsolute(path) && (path = join(process.cwd(), path))
 		if (!existsSync(path)) {
-			console.info(`[NOT_EXIST]: "${path}" does not exist`)
+			options.silent !== false && console.info(`[NOT_EXIST]: "${path}" does not exist`)
 			continue
 		}
 		// get stat
@@ -37,7 +48,10 @@ export async function rm(paths: PathLike | PathLike[], options?: RmOptions): Pro
 
 		// get files
 		const files = await promises.readdir(path)
-		await rm(files.map(name => (typeof path === 'string' ? join(path, name) : name)))
+		await rm(
+			files.map(name => (typeof path === 'string' ? join(path, name) : name)),
+			options
+		)
 
 		// remove dir
 		await promises.rmdir(path, options)
@@ -48,14 +62,18 @@ export async function rm(paths: PathLike | PathLike[], options?: RmOptions): Pro
  * remove file sync
  *
  * @param paths - path or file, support array
+ * @param options - RmOptions
  */
-export function rmSync(paths: PathLike | PathLike[], options?: RmOptions): void {
+export function rmSync(
+	paths: PathLike | PathLike[],
+	options: RmOptions = { silent: true, force: true }
+): void {
 	if (!(paths instanceof Array)) paths = ([] as PathLike[]).concat(paths)
 
 	for (let path of paths) {
 		typeof path === 'string' && !isAbsolute(path) && (path = join(process.cwd(), path))
 		if (!existsSync(path)) {
-			console.info(`[NOT_EXIST]: "${path}" does not exist`)
+			options.silent !== false && console.info(`[NOT_EXIST]: "${path}" does not exist`)
 			continue
 		}
 		// get stat
@@ -67,7 +85,10 @@ export function rmSync(paths: PathLike | PathLike[], options?: RmOptions): void 
 
 		// get files
 		const files = readdirSync(path)
-		rmSync(files.map(name => (typeof path === 'string' ? join(path, name) : name)))
+		rmSync(
+			files.map(name => (typeof path === 'string' ? join(path, name) : name)),
+			options
+		)
 
 		// remove dir
 		rmdirSync(path, options)